requirements.txt
autogen-agentchat
autogen-ext[azure]
azure-ai-projects
tiktoken
openai
python-dotenv